python_app/requirements.txt
fastapi[all]
nats-py
uvicorn
pydantic
pytz
mypy
types-aiofiles
types-pytz
